refactored main components for deployment
