+++ server/package.json
@@ -2,10 +2,9 @@
   "name": "world-messenger",
   "version": "0.0.1",
   "description": "Messenger with any language",
-  "main": "server.js",
   "scripts": {
-    "start": "node server.js",
-    "server": "nodemon server.js"
+    "start": "node ./bin/www",
+    "debug": "nodemon ./bin/www"
   },
   "author": "Team Leonard",
   "license": "MIT",
@@ -14,10 +13,13 @@
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "concurrently": "^4.1.0",
+    "cookie-parser": "^1.4.4",
+    "debug": "~2.6.9",
     "express": "^4.17.1",
     "is-empty": "^1.2.0",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.5.15",
+    "morgan": "^1.9.1",
     "multer": "^1.4.1",
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
